LPD-91502 Reorganize dashboard components
