+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/inventory/InventoryContext.tsx
@@ -5,7 +5,7 @@
 
 import React, {createContext, useCallback, useReducer} from 'react';
 
-import {initialSpace} from './components/SpacesDropdown';
+import {initialSpace} from '../common/SpacesDropdown';
 
 type Item = {
 	externalReferenceCode?: string;
