+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/inventory/components/GroupByDropdown.tsx
@@ -6,9 +6,9 @@
 import {Option, Picker} from '@clayui/core';
 import React from 'react';
 
+import PickerTrigger from '../../common/PickerTrigger';
 import {Item} from './FilterDropdown';
 import {IAllFiltersDropdown} from './InventoryAnalysisCard';
-import PickerTrigger from './PickerTrigger';
 
 const defaultStructureTypes: Item[] = [
 	{
